requirements.txt
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
filelock==3.3.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==4.0.1
idna==3.3
mysql-connector-python==8.0.26
oauthlib==3.1.1
pipenv==2021.5.29
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
